feat(repeater): 修改分支信息中的修改时间字段类型
将 BranchInfo 模型中的 modified_time 字段从 int 类型改为 float 类型，
以支持更精确的时间戳存储。

+++ Repeater/repeater/plugins/repeater_main/commands/_clients/_data_manager/_branch_info.py
@@ -5,7 +5,7 @@ class BranchInfo(BaseModel):
     """Branch Info"""
     branch_id: str = ""
     size: int = 0
-    modified_time: int = 0
+    modified_time: float = 0
     readable_size: str = ""
 
     def created_time(self) -> datetime:
